add injection, small fixes

+++ GoogleSearchImageTest/Controllers/ResultController.cs
@@ -12,7 +12,17 @@ namespace GoogleSearchImageTest.Controllers
 {
     public class ResultController : ApiController
     {
-        private GoogleSearchImageTestContext db = new GoogleSearchImageTestContext();
+        private readonly IGoogleSearchImageTestContext db = new GoogleSearchImageTestContext();
+
+        public ResultController()
+        {
+
+        }
+
+        public ResultController(IGoogleSearchImageTestContext context)
+        {
+            db = context;
+        }
 
         public SearchResult Get(int id)
         {
@@ -31,15 +41,18 @@ namespace GoogleSearchImageTest.Controllers
                 if (isUpdate)
                 {
                     db.SearchResults.Attach(searchResult);
+                    var itemResult = db.SearchResults.Include(s => s.Items).FirstOrDefault(s => s.Id == searchResult.Id);
+                    if (itemResult != null)
+                    {
+                        var deletedItems = itemResult?.Items.Where(i => i.Deleted).ToList();
+                        db.SearchResultItems.RemoveRange(deletedItems);
+                    }
                 }
                 else
                 {
+                    searchResult.Items.RemoveAll(i => i.Deleted);
                     searchResult = db.SearchResults.Add(searchResult);
                 }
-
-                var itemResult = db.SearchResults.Include(s => s.Items).First(s => s.Id == searchResult.Id);
-                var deletedItems = itemResult.Items.Where(i => i.Deleted).ToList();
-                db.SearchResultItems.RemoveRange(deletedItems);
 
                 db.SaveChanges();
             }
@@ -52,7 +65,7 @@ namespace GoogleSearchImageTest.Controllers
 
             if (isUpdate)
             {
-                var response = Request.CreateResponse(HttpStatusCode.OK); 
+                var response = Request.CreateResponse(HttpStatusCode.OK);
                 string uri = Url.Link("DefaultApi", new { searchResult.Id });
                 response.Headers.Location = new Uri(uri);
                 return response;
